Move proposal listing tab to meeting package.
Also add filter for decided state.
Also simplify inheritance structure, eliminate
one level.
